func_8003CEB8 and func_8003AD6C MATCH; 872 -> 873
func_8003CEB8: one declaration, `s16 crc` for a CRC-16.  The permuter found
it paired with a named mask that is worth nothing alone.  The four zero-init
shapes measured earlier were the wrong axis -- the residue read as a
shared-zero question and was a type question.

func_8003AD6C (the card-flip animation): five things.  The two stack locals
are an ARRAY -- gcc never puts one in a register, which is what puts them at
sp+0x10/0x14 and gives retail's six extra instructions.  The s16 narrowing
of a sum needs two statements or gcc computes the whole thing in HImode and
loads both operands with lhu.  D_8009B0D8 wants the scalar declaration plus
as -G0 (gp=0, so any -G is free), which stops the lui being hoisted into a
branch delay slot retail leaves empty.  And `p + 0xC` written INLINE at all
six sites, not through a local, is the last 21 -- three separate names match
too, one shared name does not.

+++ src/func_8003CEB8.c
@@ -1,7 +1,7 @@
 #include "common.h"
 
 s32 func_8003CEB8(u8 *arg0, s32 arg1) {
-    s32 crc;
+    s16 crc;
     s32 i;
     s32 j;
